Tree Construction: Part 1, node subsets
In the first phase, we find a super-set of the nodes in the merge tree.

We use a depth metric to determine when we've finished.
The depth is increased by the number of parents, and decreased by the number of
children.

The nodes are stored on a queue, rather than a stack. With a stack, the entire
repository would be covered before we determined the end of the current merge.

The queue allows us to grow along the master branch as well as the other commits of
the merge to determine where the merge was branched from. Once we find that point,
phase-one has finished. In the process, we've discovered the number of children of
each commit, and know a little about the topology of the repository.