Bug Fixes MergeTree:
Adds children to the tree. When things are passed to asynchronous functions, 'this'
gets changed from the object we're dealing with to the window object. It isn't what
we want. So we pass the mergetree object we're dealing with as the resulting
promise from phase 1 after adding the children mappings.

The issue with the missing nodes stems from depth hitting zero because children were
being added that were already added. Thus, it would remove depth when it shouldn't
have since the children were not all unique. This occurs when there are many ways to
access a single node. Check 'atlassian/aui'