Mergetree: Phase 1 and 2, depth limiting
We re-counted previously visited branch points, causing the depth calculations to be
incorrect. We've added the necessary stuff to make it so that nodes are not visited
more than once.

We need to ensure that we have the minimal depth of a given branch point for the
calculations and will need to adjust the depth accordingly.

Phase 2 completely removes the depth limiting, only using the node queue. Since
we've done most of the heavy lifting in phase 1, this may be sufficient for correct
tree construction. Depth limiting should be checked though and determine if that is
necessary.