Mergetree: Phase 1 and Phase 2: Depth assignments
Phase 1 minimizes the depth to be the smallest possible depth from all children of
the node.
We provide the depth as part of the node lookup table.

Phase 2 uses the depth to determine which is the parent. It should be noted that
this part is not yet correct. It selects the first one that is not less than the
depth. This works in the order in which it was visited, not the order of the
parents. We should try to find the child that is closest in depth to this node and
use that as the real parent.